lm1 ready, and organized
